consistency changes and missed lint errors, separate out mockrequest class
